rename gdb.c to debugger.c
